farm: drop fcntl lock
For a second thought, even QEMU can issue multiple requests to the same object
we are guaranteed that those requests don't overlap on each other. Recovery
code and IO code might access the same object, but the race is excluded by epoch
so we don't need per-object lock at all for now.

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>