sheep: simplify the cluster driver interface for blocking events
Let sd_block_handler handle the fine details of how to handle an incoming
blocking event.  By passing the sender node structure we can easily handle
ignoring it on other nodes, and by keeping a local operation in progress
flag in group.c we can replace the callbacked flag in the on the wire
events with a way simpler mechanism.

The only slightly complicated bit is that zk_notify_blocked in the zookeeper
backend can now go negative for a short period of time, so we explicitly
have to check for it beeing positive in two places.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>