sheep: fix error handling in epoch_log_remote
Make sure we exit if either get_latest_epoch or epoch_log_read fails.

The get_latest_epoch failure case might not be obvious, but given that
local_stat_cluster is a a force operation we might not have any epoch
log on thise node by the time it is called.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>