sheep: collie cluster recover needs to start recovery
Currenly we can easily get into a situation where we can't read objects
after losing a node in an offline cluster and then doing a manual recovery.

To fix this call start_recovery from cluster_manual_recover.  Also move
get_vnodes_from_epoch into group.c and rename it to fit with the rest of
the vnode_info functions now that is is used outside of recovery.c.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>