sheep: factor leave_list processing
Factor handling of the leave list into a few helpers to make it more obvious.
Also rename the leave name component to failed/failed_nodes to make the use
case more obvious.  Finally, remove the nodes/leave_nodes union in favor of
a single array, as an union of two fields with the same type is fairly
confusing.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>