sheep: handle new nodes during WAIT_FOR_JOIN state
There is no problem letting new nodes join a cluster in WAIT_FOR_JOIN state as
long as they aren't counted towards the quorum of nodes requires to restart
the cluster.  For this repurpose the CJ_RES_JOIN_LATER return value that is
currently equivalent to CJ_RES_FAIL to indicate we have a node that could
join, but wasn't part of the cluster before shutdown.  A new nr_delayed_nodes
counter and list of delayed nodes is also added to the join message to make
sure all nodes know about this kind of delayed node.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>