sheep: take truns to monitor EPOLLIN and EPOLLOUT events.
When both EPOLLIN and EPOLLOUT events occur, client_hander() will always
process EPOLLIN event firstly, but this is not reasonable in some case. For
example:
     Client                                     Server
        |                                        |
        |                                   monitor(EPOLLIN)
        |                                        |
      write()  >- send request pkg ->            |
        |                                      read()
        |                                        |
        |                              [1]: monitor(EPOLLIN and EPOLLOUT)
        |                                        |
        |   <- send response pkg(1/2) -<      write()
        |                                        |
        |                                     EPOLLIN occur
                                       [2]:  read()
[1]:
After the server read the request pkg from a client, it will monitor both
EPOLLIN and EPOLLOUT events. Then the server is going to write response pkg to
the client, but only 1/2 data was send to the client for network busy.

[2]:
Before all of the response pkg was send to the client, EPOLLIN occur, then the
server will call read() to this fd. In this case, the respone pkg could not be
send to the client as soon as quickly. then the server has not chance to send
the response pkg completly, it may lead to dead lock.

The simplest way to fix this problem is that, let sheep take truns to
monitor EPOLLIN and EPOLLIUT events.

Signed-off-by: Yunkai Zhang <qiushu.zyk@taobao.com>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>