sheep: free all requests when connection is dead
It's important to free all requests when connection is dead(especially when
EPOLLERR occur, some requests might be not cleared correctly), or it will lead
to memory leak and sys->nr_outstanding_reqs would be always larger than 0 as a
result sheep could not be shutdown.

Conneciont's request is created by alloc_request() and will be kept in four
places:
  1) ci->rx_req       -- after alloc_request()
  2) inflight         -- after reset ci->rx_req and queue_request()
  3) ci->dones_reqs   -- after put_request() added to ci->dones_reqs list
  4) ci->tx_req       -- after init_tx_hdr() and assigned to ci->tx_req
When request is inflight, ci->refcnt will always be larger than 0, but it will
be added to ci->dones_reqs after work_fn finished.

As long as we can promise that one request only be kept in one of these four
places, than we can free it correctly.

This patch do this clear work in clear_client().

BTW: add and update some log info

Signed-off-by: Yunkai Zhang <qiushu.zyk@taobao.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>