sheep: use stored nr_copies and flags
For a sheep that is restarted we can trust the local nr_copies and cluster
flags stored in the config file, so read it from there if is present and
only pick up the values from the join message if the sheep does not
have a configuration yet.

Also change the type of the nr_copes value in struct cluster_info so
that get_cluster_copies can be used on it without warnings.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>