sheep: always access sys->status directly
A lot of access already need to be direct, e.g. for using switch statements.
Change the remaining instances to also directly access it and remove the wrappers
for it (and sys_flag_nohalt).

sd_leave_handler also got another additional cleanup by using a switch statement
and running the nr_zones check (uselessly but cleaner) for the halt state as well
so that it matches the join side.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>