collie: rename 'collie recover' to 'collie recover force'
Rename 'collie recover' to 'collie recover force' so that we can introduce
other two recover subcommands: {enable|disable} in the next patch. In order to
keep consistent naming, I also renamed some related functions and flags.

As we need to express subcommand's subcommand(such as 'recover' is a
subcommand of 'collie' and 'force' is a subcommand of 'recover'), I have
modified subcommand's data structure and refactored some helper funcions a bit.

Signed-off-by: Yunkai Zhang <qiushu.zyk@taobao.com>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>