logger: keep the log fd private to the logger process
We only need to access the log file from the logger process, so make sure
it only is available there and not exposed to the sheep which will keep
(leak) a reference to the initial one and won't be able to use it after
is has been rotate.

Also kill the unused log to syslog feature to make the change easier.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>