sheep: add plain store driver
This introduces a storage driver 'plain' based on the current storage
interface.  The design of the plain store is similar to one of the
farm driver.  The main difference is that the current farm uses the
sha1 based backend store for stale objects, but simple_store uses a
flat directory for them.  With this design, plain store can move
objects from the working directory to the backend store efficiently
with rename(2).

Here are pros vs cons of the plain store against the current farm.

Pros:
 - faster recovery
 - smaller and simpler
 - would be a good example to introduce other storage drivers

Cons:
 - cluster snapshot is not supported
 - stale objects are not deduplicated
 - there is no sha1 verification

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>