plain_store: store stale objects to .stale/oid.epoch
This is necessary to pass test 034.

It is wrong to store stale objects without epoch info.  It is because
if the objects are updated while leaving cluster, sheep cannot detect
it and uses the local stale objects as the latest one.  It is
mandatory to store stale objects with epoch information like old farm
code or simple store.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>