net: add a send timeout for sockfd
This fixes the bug found by 035.

This is kind of a subtle case:

'sendmsg' of A  would fill its sock buffer to full whlile the other end B is failed
in the middle of the 'sendsmg'. The sock buffer of A is full, then kernel block
the thread awaiting on the buffer drain, which will never happen because B is
down. Keepalive won't be triggered because the kernel detect that sock buffer
has data to send. So the final result is that, the blocked thread will never be
woke up.

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>