recovery: fix wrong index in oid_in_recovery()
rw->done is index of the original next object to be recovered and also the
number of objects already recovered. So rw->done - 1 identify an already
recovered object. A another caller case is that, if rw->done == 0, we'll end
up with a -1 valule, which might cause seg fault in the array.

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>