ops: refactor read_copy_from_replica()
This function does competely the same as gateway_read_obj(), so we'd better
directly call gateway_read_obj() by faking a request.

We didn't call it in the old code because a nasty nested requests problem over
the same FD. Now with short thread and sockfd, we don't have this kind of
problem.

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>