sheep: add support for manual recovery
This implements disbling object recovery.  When recoery is disabled,
sheep will suspend the recovery process after it recovers objects in
the prio_oids queue.  The suspended recovery is resumed after new
objects are pushed into the prio_oids queue.  This means that
unaccessed objects are not recovered at all.

Note that sheep increments epoch even when recovery is disabled.  If
sheep receives a write request, it will place the updated object based
on the current node membership.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>