store/plain: use link to recover local objects
It is not safe to rename stale objects to the working directory.  When
reading objects for recovery, default_read reads the working directory
first, and then, searches stale objects.  However, if the objects are
renamed from the stale directory to the working directory on the node
at the same time, the recovery read request fails to find the object.

To fix this race condition, this patch uses link/2 to keep stale
objects.  This fixes test failure of tests/028.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>