sheep: notify vdi copies number before we create the vdi
We can not notify the vdi copies number to the cluster after creating
the vdi, because if a recovery happens in the time window after
cluster_vdi_add() finishes, and just before post_cluster_vdi_add()
notifies the copies number of the new vdi to all the nodes in the cluster,
it may causes screen_object_list() fail to get vdi copies number of
the objects of the newly created vdi, and we can not make it wait for
the copies number because it would cause another race condition.

Notify the copies number before we create the vdi is a good solution to
this problem, it has no harm even if the creation of vdi fails.

Signed-off-by: levin li <xingke.lwp@taobao.com>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>