sockfd cache: revalidate node after grab failure
To fix following case:

The node is deleted, but someone askes us to grab it.The nid is not in the
sockfd cache but probably it might be still alive due to broken network
connection or was just too busy to serve any request that makes other nodes
deleted it from the sockfd cache. In such cases, we need to add it back.

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>