zookeeper: fix spurious seq node events
For error conditions, such as operation timeout, the write of seq node will fail.
We couldn't simply rewrite it because this would result in the same event to be
written multiple times. Sheep can't handle this situation at all.

We must panic out for error conditions for writing seq node.

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>