sheep: better error handling of numbers specified in command line arguments
Current sheep treats this command line option as a valid port number:
"-p 7000/store/sheep". But I believe this should be treated as an
error. This sort of option can be passed if users forget to type ' '
between port number and path for sheep directory.

This confusing behaviour can be found in handling port number, log
level, zone ID, and free space size. This patch fixes it.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>