logger: format in logger process
Current sheep formats log lines in sheep process before sending them
to logger process. This patch moves the formatting from sheep to
logger because it will make a noise for workload of sheep (of course
they are sublte but it is better to do this).

In addition, this patch let logger obtain timestamp with
gettimeofday() instead of time(). The purpose of this change is more
accurate timestamp. Because many logs can be generated in one second,
second is not suitable unit of timestamp for collecting multiple log
files and sorting them with timestamp.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>