logger: rename and change a role of select_log_formatter()
This patch renames select_log_formatter(). The new name is
early_log_init().

early_log_init() is also used for selecting log formatter, but it
also used for registering information of logger user. The information
is registered as 2nd parameter of the function which is typed struct
logger_user_info.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>