sheep: don't read epoch log from local in epoch_log_read_remote
This patch uses get_vnode_info to get the current node membership in
epoch_log_read_remote.  It is because the delayed nodes don't have
epoch log in local store.  This also moves epoch_log_read_remote to
group.c since it is not a function to access local store.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>