md: add dummy md proper
This add the dummy objects mapping code, which use consistent hashing, for md
like support which aims to manage multiple disks in one node.

This patch just mainly add the compile support for md.c and some initilization
code modifcation to adopt to later series in the sheep core code.

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>