sheep: make epoch_log_read() clean
Current epoch_log_read() and the way of treating epoch files are
confusing. Because they heavily depend on the condition:
sizeof(time_t) < sizeof(struct sd_node). The condition is true
currently, but this requirement is implicit and hard to understand.

In addition, epoch_log_read() doesn't handle a rise of EINTR when it
calls read(). This behavior might cause a subtle bug.

This patch eliminates the implicit requirement and adds the handling
of EINTR. For doing it, this patch implements a new function,
epoch_log_read_with_timestamp(). If callers require a timestamp at
tails of epoch files, they should call this one instead of
epoch_log_read().

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>