shepherd: a new cluster manager specialized for sheepdog
This patch adds shepherd, a new cluster manager for sheepdog. shepherd
is not a general cluster manager software. The design is heavily
depends on the internal knowledge of sheepdog. So the usage will be
limited but it can be optimized only for sheepdog.

Currently, shepherd cluster manager is designed as master-slave. So it
can be a SPOF. But I'm planning to let it be redundant by utilizing
corosync. If it can be improved, it can be an alternative of the
group management part of ZooKeeper, and it doesn't depend on JVM.

This patch adds the new directory shepherd/ for it, and the
executable binary will be produced as shepherd/shepherd.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>