sheep: clean up SD_RES_xxx macros
Currently it is a little bit mix that we put SD_RES_xxx for collie
in sheepdog_proto.h but we put some SD_OP_xxx for collie in internal_proto.h.

Let's take a uniform view that take both inter-sheep and collie-sheep as
internal protocols. In this way, collie and QEMU won't share the same number
space to interfere with each other (QEMU is unaware of collie at all).

So the final number space is:

QEMU - sheep: use 1 to 80(inclusive) of number space
inter-sheep and collie - sheep: above 80 of number space

Signed-off-by: Liu Yuan <tailai.ly@taobao.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>