employ gnu99 style of GCC
This patch lets sheepdog employ gnu99 (C99 + gnu extensions) style of
GCC. The main benefit of gnu99 is allowing interleaved statements and
declarations. sheepdog source tree has many functions with lots of
local variable declarations on their head part. They are harmful from
the perspective of readability. Basically, variable declarations
should be delayed until they are actually required. With gnu99, we can
write interleaved statements and declarations, so this can improve the
readability of sheepdog.

The RB_ROOT constant is not friendly with gnu99, so this patch also
add a minor change to it. In addition, RB_ROOT cannot be used in
statements (not declarations). This patch also prepares a new macro
INIT_RB_ROOT() for such usecases.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>