sheep: get vdi bitmap from joined node
If we stop sheepdog without the shutdown command, the number of nodes
will be rapidly reduced and it will easily happen that all the
replicas are lost.  In such case, it is better to get vdi bitmap from
the newly joined node too so that we can recover so many vdis as we
can.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <tailai.ly@taobao.com>