sheep: remove '-s' option
We use '-s' option as a hint of weighting. This actually doesn's stop sheep from
using extar space that '-s' specifies. Ask users to give weight hint is
nonsense if we can weight the node transparently.

'-s' would be meaningless too when we add reweight support.

This is a prepare patch for reweight patch.

Signed-off-by: Liu Yuan <namei.unix@gmail.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>