recovery: fix tgt_epoch chaos
tgt_epoch is the epoch of highest epoch from where we try to recover object. We
can't simply assign it as the sys->epoch. This won't cause fatal problem because
we'll tgt_epoch--, but will cause unnecessary recovery requests that mostly fail

This patch adds a new field tgt_epoch to mean exactly it.

Signed-off-by: Liu Yuan <namei.unix@gmail.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>