sheep: handle node failure during atomic_create_and_write
If sheep fails after creating a temporary file and before renaming it
in atomic_create_and_write(), sheep cannot create the file any more
since no one doesn't clean up the temporary file.  This patch adds a
force option to atomic_create_and_write() and cleans up the temporary
file when it is enabled.

It is okay for write_config() and update_epoch_log_() to set the force
option because these function cannot be called from multiple threads
at the same time.  However, md_move_object() can be called from
multiple threads, so this patch keeps the function's behavior as it
was.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>