import debian directory maintained by debian package team
This copies the debian patches to build a deb package from
  http://patch-tracker.debian.org/package/sheepdog

The changelog files are not imported since they are auto-generated by
the next patch, and the translation files are not imported since it's
difficult for our community to maintain them.  Some style problems
detected by our checkpatch.pl are also fixed.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>