collie: add a new option --progress to "node recovery" for showing recovery progress
This patch adds a new option --progress (or -P) to the node recovery
subcommand. With this subcommand, users can show a progress of
recovery process.

Example:
$ sudo collie node recovery --progress
 99.7 % [==============================================>] 7047 / 7068

The denominator (7068 in the above case) indicates a number of entire
object which should be checked. The numerator (7047 in the above case)
indicates a number of objects which is already checked or copied.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>