sheep: remove SD_STATUS_HALT
SD_STATUS_HALT is used to avoid reading an old data when there are too
few nodes and node failure happens with them.  However, the current
implementation already allows users to read the old data when all the
replicas are lost.  It doesn't make sense to have a halt status as a
special case.

This patch removes SD_STATUS_HALT and "-m (unsafe|quorum)" option from
the collie command.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>