cache: correct block size calculation for inode object
inode and data object have different sizes and thus cache block size is
different. Current code fail to get the right size for inode object when
its last bit of bmap is set with the fixed block size.

This patch tries to get the dynamic block size based on the oid.

Signed-off-by: Liu Yuan <namei.unix@gmail.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>