trace: create a caller list in tracer_init()
In tracer_enable(), we replace all the mcount() calls in the caller
list with trace_caller().  However, we create the caller list
dynamically; we update the list when sheep calls a new function which
is not in the list for the first time.  This means that we cannot
trace the functions which were not called before we call
tracer_enable().

To include all the function into the tracer targets, this patch
creates a complete caller list just after sheep starts up.  This
achieves it with BFD (a part of binutils-dev) and makes the current
tracer code much simpler.

This can also remove trace.ld, and allows us to see a line number when
debugging with GDB.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>