add notrace to only entry point and exit point
It is difficult to know which function needs a notrace directive.  We
are using notrace to avoid starting tracer recursively while tracing,
but we can do the similar thing by using a thread local variable.
With this patch, notrace is necessary only to the entry point
(trace_function_enter) and the exit point (trace_function_exit).  This
patch removes notrace compiler.h so that we don't abuse the directive
any more.

This patch also allows us to use tracer with debug build.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>