sheep: add a checker for thread type
This introduces a two maker "main_fn" and "worker_fn".  If the
function must be called in the main thread, add "main_fn" to it.  If
the function must be called in the worker thread, add "worker_fn" to
it.  The markers are used only when we enable tracer infrastructure.
The check is done at the entry point of the function.

Adding the marker is simpler and easier than writing assert.  In
addition, programmers can easily understand which function must be
called in the main/worker thread.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>