corosync: show better error messages
If cpg_initialize() returns CS_ERR_SECURITY, the error reason is sheep
doesn't have a privilege to use corosync because of corosync
configuration.  In this case, printing "Permission denied" is better.

In the other cases, corosync is actually running and there is another
error reason.  This removes "is corosync running?" from the error
message in the default cases.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>