sheep: make the vid deletion proceduer correct order
Current VDI deletion process deletes VDI in this order:
1. delete objects which belong to the VDI
2. notify deletion of the VDI for removing object list cache
3. clear inode object with zero

But this should be like this:
1. delete objects which belong to the VDI
2. clear inode object with zero
3. notify deletion of the VDI for removing object list cache

Because in the process of removing object list cache, existing of the VDI is
checked with the name of the VDI is zero-cleared or not
(objlist_deletion_work()). So objlist_deletion_work() would have a possibility
of leaving the object list cache incorrectly. This patch makes the process
correct order.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>